requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
